requirements.txt
flask
youtube-transcript-api
google-generativeai
yt-dlp
